# docs/swagger.yaml
definitions:
...
  controllers.DeleteResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo:
        $ref: '#/definitions/models.Photo'
      photoID:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.Photo:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.SocialMedia:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.User:
    properties:
      age:
        type: integer
      ban_reason:
        type: string
      banned_at:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      last_seen_at:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Photo'
        type: array
      role:
        type: string
      social_medias:
        items:
          $ref: '#/definitions/models.SocialMedia'
        type: array
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/comments:
    get:
      consumes:
      - application/json
      description: Get all comments in finalproject
      produces:
      - application/json
      responses:
        "200":
          description: Get all comments success
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all comments
      tags:
      - comment
  /api/v1/comments/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete comment identified by given ID
      parameters:
      - description: ID of the comment
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete comment success
          schema:
            $ref: '#/definitions/controllers.DeleteResponse'
        "400":
          description: Invalid comment id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: Get comments identified by given id
      parameters:
      - description: ID of the comments
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get comment success
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid comment id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get comments
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update comment identified by given id
      parameters:
      - description: ID of the comment
        in: path
        name: commentId
        required: true
        type: integer
      - description: Comment update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: Update comment success
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - comment
...
  /api/v1/photos/{photoId}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for photo with given id
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get all comments success
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "400":
          description: Invalid photo id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all comments for specific photo
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create comments for photo identified by given id
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      - description: Comment request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "201":
          description: Create comments success
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create comments
      tags:
      - comments
...
  /comments/create/{photoId}:
    post:
      consumes:
      - application/json
      description: Create comments for photo identified by given id
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      - description: Comment request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "201":
          description: Create comments success
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create comments
      tags:
      - comments
  /comments/delete/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete comment identified by given ID
      parameters:
      - description: ID of the comment
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete comment success
          schema:
            $ref: '#/definitions/controllers.DeleteResponse'
        "400":
          description: Invalid comment id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - comment
  /comments/get/{commentId}:
    get:
      consumes:
      - application/json
      description: Get comments identified by given id
      parameters:
      - description: ID of the comments
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get comment success
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid comment id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get comments
      tags:
      - comments
  /comments/getall:
    get:
      consumes:
      - application/json
      description: Get all comments in finalproject
      produces:
      - application/json
      responses:
        "200":
          description: Get all comments success
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all comments
      tags:
      - comment
  /comments/getall/{photoId}:
    get:
      consumes:
      - application/json
      description: Get all comments for photo with given id
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get all comments success
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "400":
          description: Invalid photo id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all comments for specific photo
      tags:
      - comment
  /comments/update/{commentId}:
    put:
      consumes:
      - application/json
      description: Update comment identified by given id
      parameters:
      - description: ID of the comment
        in: path
        name: commentId
        required: true
        type: integer
      - description: Comment update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: Update comment success
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - comment
...
swagger: "2.0"